pyproject.toml
[project]
name = "podcast_download"
authors = [
{ name = "Enucatl", email = "1423701+Enucatl@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">= 3.9"
dependencies = [
    "click",
    "tqdm",
    "requests",
    "feedparser",
    "python-slugify",
    "pydub",
]
dynamic = ["version"]

[build-system]
requires = [
    "setuptools >= 61",
    "wheel",
    "setuptools-git-versioning",
]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "ruff",
]

[tools.setuptools-git-versioning]
enabled = true

[tool.ruff]
select = ["ALL"]

[project.scripts]
download_podcast = "download:main"
